requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.1
django-admin-sortable2==0.7
django-appconf==1.0.2
django-classy-tags==0.8.0
django-colorful==1.3
django-compressor==2.2
django-cookie-law==2.0.1
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-htmlmin==0.10.0
django-meta==1.4.1
django-nested-inline==0.3.7
django-redis-cache==1.8.1
django-taggit==0.23.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
firebase-admin==2.16.0
gcloud==0.17.0
get==1.0.3
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
gunicorn==19.9.0
hiredis==1.0.0
html5lib==1.0.1
httplib2==0.12.1
idna==2.7
jws==0.1.3
lxml==4.2.5
M2Crypto==0.30.1
msgpack==0.6.1
mysqlclient==1.3.13
oauth2client==3.0.0
Pillow==5.2.0
ply==3.11
post==1.0.2
protobuf==3.7.1
public==1.0.3
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.4.3
Pyrebase==3.0.27
pytest-runner==4.2
python-jwt==2.0.1
pytz==2018.5
query-string==1.0.2
rcssmin==1.0.6
redis==2.10.6
request==1.0.2
requests==2.21.0
requests-toolbelt==0.7.0
rjsmin==1.0.12
rsa==4.0
six==1.11.0
slimit==0.8.1
soupsieve==1.7.3
sqlparse==0.2.4
urllib3==1.23
webencodings==0.5.1
